Update on changes only

+++ update.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-rm -rf _site/
-git pull -q
-bundle install
-git checkout Gemfile.lock
-bundle exec jekyll build
+OLD_HASH=$(git rev-parse --short HEAD)
+git pull
+if [[ "${OLD_HASH}" != $(git rev-parse HEAD) ]]; then
+    rm -rf _site/
+    bundle install
+    git checkout Gemfile.lock
+    bundle exec jekyll build
+fi
